Literasee: Assessment Literacy Content Development via R

Assessment Literacy for R.

Details

Package:

Literasee

Type:

Package

Version:

0.0-0.6

Date:

2017-5-12

License:

GPL-3

LazyLoad:

yes

Literasee is an experimental package for the R software environment designed to ease in the creation of assessment literacy related content.

References

Betebenner, D. W. (2009). Norm- and criterion-referenced student growth. Educational Measurement: Issues and Practice, 28(4):42-51.