Places figure and caption appropriately within the supplied file.

Places figure appropriately within file with caption.

placeFigure(
    files,
    rows=1,
    columns=1,
    figure.id=NULL,
    caption=NULL,
    caption.position="top",
    pdf.width=NULL,
    html.width=NULL,
    page.break=FALSE,
    normalize.path=FALSE
  )

Arguments

files
Character string of files in which figures are placed. Multiple files will be put in a rows x columns grid by row (i.e. with four files in a 2x2 grid, the first two files will be the top row and the last two files will be the bottom row.)

rows
Postive integer, defaults to 1.

columns
Positive integer defining number of columsns, defaults to 1.

figure.id
Figure identifier.

caption
Figure caption.

caption.position
Figure caption position. Either ‘top’ (the default) or ‘bottom’.

pdf.width
The width of the PDF output.

html.width
The width of the HTML output (if single image in the figure - ignored for multiple image grid).

page.break
Boolean variable indicating whether there should be page breaks (defaults to FALSE).

normalize.path
Boolean variable indicating whether the should paths in files should be normalized with normalizePath (defaults to FALSE).