Function for automatic technical report generation from SGP analyses.

The function reportSGP is used to create automatic or customized technical reports associated with SGP analyses using the SGP object produced as part of the SGP analysis.

reportSGP(sgp_object=NULL)

Arguments

sgp_object
An SGP object resulting from SGP analyses.

Value

Returns a technical in multiple output formats including HTML and PDF.

Examples

    reportSGP()

#> [1] "This function will (eventually) produce automatic reports based upon SGP analyses!"