Render R Markdown in Multiple Output Formats

Render the input file to the specified output formats using pandoc. If the input requires knitting then knit is called prior to pandoc.

renderMultiDocument(rmd_input,
  report_format = c("HTML", "EPUB", "PDF"),
  cleanup_aux_files = TRUE,
  number_sections = TRUE,
  number_section_depth=3,
  convert_header_levels=c(5,6),
  toc = TRUE,
  toc_depth = 2,
  self_contained = TRUE,
  docx_self_contained = FALSE,
  dev="png",
  literasee_template = "sgp_report",
  html_template = "default",
  epub_template = "default",
  pdf_template = "default",
  html_css = "default",
  epub_css = "default",
  docx_css = "default",
  cover_img = NULL,
  add_cover_title = FALSE,
  bibliography = "default",
  csl = "default",
  pandoc_args = NULL,
  ...)

Arguments

rmd_input
Input file. Must be an R Markdown (.Rmd) file located in the working directory).

report_format
Report output format(s) to convert Rmd file to. Pass a vector of format names to render multiple formats. Defaults to the three polished versions (c("HTML", "EPUB", "PDF")). "DOCX" is also available, but is intended to be a draft-version document as it does not support many of the stylization features included in the other formats.

cleanup_aux_files
Boolean. Defaults to TRUE to delete the intermediate markdown files and directories be upon successful completion.

number_sections
Boolean. Should sections, subsections, and subsubsections be numbered. Default TRUE

number_section_depth
Numeric. Depth of headers to include in section numbering. Default is 3 (subsubsections).

convert_header_levels
Numeric. Markdown header levels 5 and 6 can be used as figure captions in HTML document. These headers are converted back to in image captions for PDF format.

toc
Boolean. TRUE to include a table of contents in the output

toc_depth
Numeric. Depth of headers to include in table of contents. Default is 2 (subsections).

self_contained
Produce a standalone HTML file with no external dependencies, using data: URIs to incorporate the contents of linked scripts, stylesheets, images, and videos.

docx_self_contained
Produce a standalone DOCX file with no external dependencies. Allows copy and paste of HTML content to Word document without serving the local directory.

dev
Graphics device to use for figure output (defaults to png).

literasee_template
Name of Literasee package style templates to use for rendering. Default is "sgp_report" (currently the only option available). Located at system.file('rmarkdown', 'templates', 'sgp_report', 'resources', package = 'Literasee'). Note that alternative templates must adopt the same naming conventions ('html_report.html', 'pdf_report.tex', etc).

html_template
Pandoc template to use for HTML rendering. Pass "default" to use the Literasee package default template. Note that if you don't use the "default" template then the features of multi_document won't be available. Additionally, there is a 'simple' version that does not include the table of contents or the two-toned background (useful for printing/PDF via browser).

epub_template
Pandoc template to use for EPUB rendering. The "default" uses Pandoc's base template.

pdf_template
Pandoc template to use for PDF rendering. Pass "default" to use the Literasee package default template.

html_css
The CSS for HTML format if other than the default within the package.

epub_css
The CSS for EPUB format if other than the default within the package.

docx_css
The CSS for DOCX format if other than the default within the Grmd package.

cover_img
Directory path for image to use for cover to EPUB document.

add_cover_title
EXPERIMENTAL. Add the document title and subtitle to the EPUB cover image. Currently set up for only .jpg files and tailored to an image with dimensions roughly 2100 x 2800.

bibliography
The LaTeX style bibliography file (.bib, .bibtex, etc.) to include. The default file included in the package includes references often used by the author and will likely be insufficient for most users.

csl
The CSL (citation style language) to use to format bibliography (if included). Defaults to APA 5th edition. See this Github repository for alternatives. If NULL pandoc's defaults will be used.

pandoc_args
Additional command line options to pass to pandoc

...
Passed ontomulti_document and/or html_document.

Value

The compiled documents are written into the working directory. Additional directories are created for each of the output formats.

Details

This function is a wrapper function that combines calls to render and direct system calls to Pandoc. The function begins with a rendering of the HTML version of the document and saves the markdown file from this rendering and uses that as a base from which other formats can be rendered. Other formats require specific post-processing, which is also accomplished in the function.

See also

Rmarkdown knit, render, pandoc

Examples

not_run({
  library(Literasee)

  # Render all four formats.
  # Use --webtex to process LaTeX style math equations.

  renderMultiDocument("My_Report.Rmd",
                      report_format = c("HTML", "EPUB", "PDF", "DOCX"),
                      cleanup_aux_files = FALSE,
                      pandoc_args = "--webtex")

})